requirements.txt
tensorflow
scikit-learn
django
numpy
pandas
xgboost
djangorestframework
